fix compilation error on 64bits
git-svn-id: https://svn.code.sf.net/p/kaldi/code/sandbox/karel@686 5e6a8d80-dfce-4ca6-a32a-6e07a63d50c8

+++ src/cudamatrix/cu-device.cc
@@ -54,7 +54,7 @@ void CuDevice::PrintProfile() {
 
 
 std::string CuDevice::GetFreeMemory() {
-  unsigned int mem_free, mem_total;
+  size_t mem_free, mem_total;
   cuMemGetInfo(&mem_free, &mem_total);
   std::ostringstream os;
   os << "Free:" << mem_free/(1024*1024) << "MB "
